broke out fn defs and classes, added helpers, rewrote add-advice
we added a condition for symbols not denoting an advisable function in
the provided database.
broke out the calss definition into individual before, after, and
around advice classes. (need to trim inner-* slots from these)
added the macro with-inner-advice, which gets the inner advice from an
advice function
added macro make-advisable, which creates an advisable-function object
with a dispatch function and the main function in the provided hash
table.
added macro defadvisable, which wraps up defun and make-advisable into
a single call.
added function pop-advice, which unwraps the advice specified for the
function specified.
broke all advice function generation into individual functions,
which cleans up the case statements by quite a bit.
commented out the add-advice macro and added a new one to use our
broken out functions.

how advice is made:
one can now add an arbitrary ammount of advice of a specific type to a
function. multiple advice can be thought of as a stack of sorts - we
define a single advice function as our base, and if any advice is
added to that, we wrap the original function up in our newly generated
advice function. its like algebra, or (american) football - to get to
the original device we have to peel/pop off the most recently defined
advice with the function pop-advice. we can use the macro
with-inner-advice to get the inner advice (second on the advice
"stack") without popping it off of the advisable-function object.